+++ CHANGES.md
@@ -4,11 +4,15 @@
 
 
 # CURRENT:
+	Version 0.43:
+	  * Stopped the bone orientation from being applied to animations. 
+	  * Improved tolerance for DAE/GLB/GLTF file imports.
+	  
+          ~~ ~~ ~~ ~~ ~~ ~~ ~~ ~~ ~~ ~~ 	  
 	Version 0.41:
 	  * Added mass metadata checker; optionally outputs to JSON with status (has armature,  has mesh, etc), internal filename, GR2Tag and local filepath. 
 		Currently uses a hardcoded file list. Will add folders later for collecting data en masse, but after the UI panel is done.
 
-          ~~ ~~ ~~ ~~ ~~ ~~ ~~ ~~ ~~ ~~ 
 	Version 0.4:
 	  * Simple GR2/DAE/GLB/GLTF works
 	  * Basic animation imports works
@@ -22,11 +26,13 @@
 
 	Version 0.3.2:
 	  * Fixed some of the CLI issues with inconsistent filenames.
-	  
-
 
 # TODO:
 	As of 21/10/25:
+	  * Add a database lookup for anim/skel combinations. If not 'GR2Tag', can just use the filename or potentially internal filename. (Probably the former more than the latter.)
+	  * Set up a prefs json for current filepaths once the UI is in.
+	  * Add the option to batch import animations from a folder or list.
+	  * If I pull the duration/framerate from the metadata, I could set it to have the frames match the imported anim. Wouldn't need it that often but might be interesting...
 	  * Find out how "GR2Tag" applies in the LSF files, see if it can be used for assigning correct skeleton file to anim files.
 	  * Automate parentage of imported animations to mesh/obj (optional)
 	  * Test for 0-size generated files. Currently they're discovered with the metadata checker, but return the same failure error as viable non-GR2 inputs.